dang sua

+++ src/main/java/com/nahuannghia/shopnhn/model/User.java
@@ -1,12 +1,20 @@
 package com.nahuannghia.shopnhn.model;
 
+import java.time.LocalDateTime;
+
 import com.nahuannghia.shopnhn.enums.UserRole;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import jakarta.persistence.*;
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "users")
@@ -18,7 +26,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "userId")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "username", nullable = false, length = 100)
     private String username;
